[feat/#14] Floating Button 컴파운드 패턴
