--- src/components/FloatingButton/CircleFloatingButton/CircleFloatingButton.module.scss
@@ -1,60 +1,0 @@
-@use 'src/styles/color' as color;
-
-.circle-floating-button {
-  width: auto;
-  height: 100%;
-  padding: 14px 16px 14px 12px;
-  border-radius: 100px;
-  border: 0;
-  background-color: color.$Primary-Orange-300;
-  box-shadow:
-    0 0 4px 0 rgb(60 37 37 / 8%),
-    0 4px 8px 0 rgb(0 0 0 / 8%),
-    0 6px 12px 0 rgb(0 0 0 / 12%);
-  display: flex;
-  align-items: center;
-  z-index: 9999;
-
-  &.small {
-    height: auto;
-    padding: 14px 18px 14px 16px;
-  }
-
-  &:hover {
-    background-color: color.$Primary-Orange-200;
-  }
-
-  &:active {
-    background-color: color.$Primary-Orange-100;
-  }
-}
-
-.bookmark {
-  background-color: color.$Grayscale-50;
-  width: auto;
-  height: 100%;
-  padding: 16px;
-  border-radius: 100px;
-  border: 0;
-  box-shadow:
-    0 0 4px 0 rgb(60 37 37 / 8%),
-    0 4px 8px 0 rgb(0 0 0 / 8%),
-    0 6px 12px 0 rgb(0 0 0 / 12%);
-  display: flex;
-  align-items: center;
-
-  &:hover {
-    background-color: color.$Primary-Orange-50;
-  }
-
-  &:active {
-    background-color: color.$Primary-Orange-50;
-  }
-}
-
-@media screen and (width <= 768px) {
-  .circle-floating-button-img {
-    width: 24px;
-    height: 24px;
-  }
-}
