+++ src/components/FloatingButton/FloatingButton.module.scss
@@ -1,0 +1,61 @@
+@use 'src/styles/color' as color;
+
+.floating-button {
+  width: 64px;
+  height: 64px;
+  border-radius: 9999px;
+  border: 0;
+  background-color: color.$Primary-Orange-300;
+  box-shadow:
+    0 0 4px 0 rgb(60 37 37 / 8%),
+    0 4px 8px 0 rgb(0 0 0 / 8%),
+    0 6px 12px 0 rgb(0 0 0 / 12%);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  &:hover {
+    background-color: color.$Primary-Orange-200;
+  }
+
+  &:active {
+    background-color: color.$Primary-Orange-100;
+  }
+}
+
+.with-children {
+  width: 140px;
+}
+
+.small-with-children {
+  width: 120px;
+  height: 54px;
+}
+
+.small {
+  width: 54px;
+  height: 54px;
+}
+
+.bookmark {
+  background-color: color.$Grayscale-50;
+}
+
+.small-img {
+  width: 24px;
+  height: 24px;
+}
+
+.small-text {
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 26px;
+  color: color.$Grayscale-50;
+}
+
+.medium-text {
+  font-size: 20px;
+  font-weight: 600;
+  line-height: 32px;
+  color: color.$Grayscale-50;
+}
