--- src/components/FloatingButton/OvalFloatingButton/OvalFloatingButton.module.scss
@@ -1,49 +1,0 @@
-@use 'src/styles/color' as color;
-
-.oval-floating-button {
-  width: auto;
-  height: 64px;
-  padding: 14px 16px 14px 12px;
-  border-radius: 100px;
-  border: 0;
-  background-color: color.$Primary-Orange-300;
-  box-shadow:
-    0 0 4px 0 rgb(60 37 37 / 8%),
-    0 4px 8px 0 rgb(0 0 0 / 8%),
-    0 6px 12px 0 rgb(0 0 0 / 12%);
-  display: flex;
-  align-items: center;
-  z-index: 9999;
-
-  &.small {
-    height: auto;
-    padding: 14px 18px 14px 16px;
-  }
-
-  &:hover {
-    background-color: color.$Primary-Orange-200;
-  }
-
-  &:active {
-    background-color: color.$Primary-Orange-100;
-  }
-}
-
-.create-form {
-  font-size: 20px;
-  font-weight: 600;
-  line-height: 32px;
-  color: color.$Grayscale-50;
-}
-
-@media screen and (width <= 768px) {
-  .create-form {
-    font-size: 16px;
-    line-height: 26px;
-  }
-
-  .plusform-img {
-    width: 24px;
-    height: 24px;
-  }
-}
